backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
